chore: theme adaptations and more responsive layouts
