+++ apps/frontend/src/features/messages/Messages.tsx
@@ -6,7 +6,6 @@ import { Message } from "./MessagesPresets";
 import { MessagingContext } from "./MessagingContext";
 import { config } from  '../../config';
 import { ProfileContext } from "../profiling/ProfileContext";
-import { date } from 'drizzle-orm/pg-core';
 
 const style = {
     width: '70%',
